Move logic to locate server around
